requirements.txt
torch==2.2.0
torchvision==0.17.0
pytorch-lightning==2.1.0 
pandas==2.0.3 
pillow==9.1.1 
einops==0.6.1 
tqdm==4.66.1   
matplotlib==3.7.2 